CORE-19951 `ExceptionCriteria` location
